chore(tracing): let a writer hold multiple clients (#5531)
This change creates an abstraction over endpoint/encoder pairs to which
`HTTPWriter`s can send payloads. Instead of holding a single `_endpoint`
and a single `_encoder`, an `HTTPWriter` now holds a list of `_client`s.
A `WriterClientBase` contains an encoder and an endpoint.

This pull request solves a problem presented by
https://github.com/DataDog/dd-trace-py/pull/5522: the
`CIVisibilityWriter` needs to send differently-encoded payloads to
different endpoints from a single trace. The solution of
`WriterClientBase` enables this while imposing as few changes as
possible on other parts of the code. For example, the `Tracer` still
only holds one `TraceWriter`.

Existing tests are enough to cover the risks of these changes.

### Changes

* Creates a new module `ddtrace/internal/writer/` containing `writer.py`
and a new file implementing the `WriterClientBase` classes
* Updates `HTTPWriter` to hold multiple clients
* Provides properties `HTTPWriter._encoder` and `HTTPWriter._endpoint`
for backward compatibility with tests that access these private
attributes
* Updates writer tests to account for multiple clients per writer